[Fix] 내 서재 파일 분리 - 오류 수정 완료
